backend/requirements.txt
# Flask Core
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Mail==0.9.1

# Authentication & Security
Flask-Security-Too==5.3.2
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.0
email-validator==2.1.0
bcrypt==4.1.2

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Supabase Integration
supabase==2.3.0
postgrest==0.13.2

# API & HTTP
requests==2.31.0
urllib3==2.1.0

# Data Processing
pandas==2.1.4
numpy==1.25.2
python-dateutil==2.8.2

# Environment & Configuration
python-dotenv==1.0.0
click==8.1.7

# Caching & Sessions
redis==5.0.1
Flask-Session==0.5.0

# Validation & Serialization
marshmallow==3.20.2
marshmallow-sqlalchemy==0.29.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Development
flask-shell-ipython==0.5.3
Werkzeug==3.0.1

# Utilities
pytz==2023.3
holidays==0.37
Pillow==10.1.0
